Calculate true total time with perf
For perf, the input includes complete call chains for every single
sample, so it is not necessary to estimate the total times
heuristically.  Instead, total time is simply the proportion of samples
for which the function was either executing or was found on the stack.
This is the approach suggested by Mike Dunlavey at
stackoverflow.com/questions/1777556/alternatives-to-gprof.