Show full filename in tooltips for Python profiles
Only works with some graphviz output formats (e.g. SVG).

Only supports the 'pstats' input format (mostly because I'm unfamiliar
with the other formats, and because I don't know if the problem is as
pressing with those as it is with Python profiles).

Fixes https://github.com/jrfonseca/gprof2dot/issues/14.