Allow compatibility with systems that do not have a unicode codec
Use a temporary file and attempt to write a unicode string into it.
If there is an error, set fallback options for the unicode strings.
Also wrap unichr to do a hex ascii string, just to be safe.

To test this behaviour:
Use a Docker image based on debian:stretch with python3 installed.